MAINT: Move mock QSBackend to conftest
